Move the project's link from the 'Try It Out!' button to the project's name and image
The 'Try It Out!' button was removed. Now either click the project's name or the Phone svg with the project's image to go to the project's link.

To keep .name centered when it is multiple lines, .name had to be set to 'display: block' because .name is now an 'a tag' which is 'display: inline'