Fix page jumping when screen is first tapped on Chrome mobile and Edge mobile
On Google Chrome mobile and Microsoft Edge mobile, the page would be longer (window.innerHeight value would be higher) than it should be. Then whenever the page is tapped, the page length/height would go to how it should be. This caused the page to jump. This would only happen on new tabs. It would not happen if the page was refreshed. This is caused by the mobile browser including the height of the keyboard when calculating window.innerHeight.